creating 1920 rays in rays array, but only the first one which is player ang minus 30 getting printed, and probably fed

+++ cub3D/inc/cub3D.h
@@ -6,7 +6,7 @@
 /*   By: erosas-c <erosas-c@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/22 14:14:54 by aaespino          #+#    #+#             */
-/*   Updated: 2024/05/22 17:26:00 by erosas-c         ###   ########.fr       */
+/*   Updated: 2024/05/22 18:59:36 by erosas-c         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -123,7 +123,7 @@ typedef struct s_player
 {
 	float	x;
 	float	y;
-	int		dir; //era p_dir pero crec que no te sentit, no cal la p i nomes alalrga els caracters que calen
+	float	dir; //era p_dir pero crec que no te sentit, no cal la p i nomes alalrga els caracters que calen
 	float	dx; //p_dir_x_cos;  //seria dx suposo (el desplacament d'x)
 	float	dy; //p_dir_y_sin; //idem per a y
 	float	n[2];
